refactor: split common and test
